# perf: use single dynamic import with dynamic code generation