- fix: load alias resolved name
- feat: bundle dynamic imports to single one

(I worked based on https://github.com/Shougo/ddc.vim/pull/70)

Measurement with capture software with using 12 sources + 3 filters,

main: 3.415s (import itself: sum 3000ms)
this: 2.085s (import itself: 700ms + 150ms)

but I think you can make it faster with loading all sources at once.

references:

- https://github.com/denoland/deno/issues/12519