@Shougo 2021-10-30
> but I think you can make it faster with loading all sources at once.

What do you mean??