@LumaKernel 2021-10-30
> What do you mean??

I can't follow the all flow for now but the dynamic import is now called twice or more with this code, so we may decrease number of calling.