@Shougo 2021-10-30
I have fixed it in #73.